% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPositionBySymbol}
\alias{getPositionBySymbol}
\title{retrive genome coordination by gene name}
\description{
This method search genome region by gene name.
}
\examples{
genome.ver  <- 'hg19'
zoom.power  <- 1
smart2.bam.files   <- c(
  "~/Dropbox_RIKEN/Public_ACCCBiT/Data/Smart-Seq2/bam/Smart-Seq2_01.bam",
  "~/Dropbox_RIKEN/Public_ACCCBiT/Data/Smart-Seq2/bam/Smart-Seq2_02.bam"
)

gb.smart2 <- orenogb2$new(
  genome.ver  = genome.ver,
  zoom.power  = zoom.power,
  bam.files   = smart2.bam.files
)
gb.smart2$getPositionBySymbol('CDK2')
gb.smart2$plotgb()
}
\keyword{search}

