% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{orenogb2-class}
\alias{orenogb2}
\alias{orenogb2-class}
\title{Class orenogb2}
\description{
orenogb2 Class (description)
}
\details{
\code{orenogb2} class is a reference class (details).
}
\section{Fields}{

\itemize{
  \item genome.ver, mm10, hg19 etc.
  \item chr. chr1, chrX etc.
  \item start.bp, genomic coordination (bp)
  \item end.bp, genomic coordination (bp)
  \item zoom.power, 1, 1/100 etc.
  \item bam.files (test01.bam,test02.bam)
}
}

\section{Contains}{

NULL
}

\section{Methods}{

\itemize{
  \item plotgb
  \item getPositionBySymbol
}
}
\keyword{documentation}

