% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotgb}
\alias{plotgb}
\title{draw a genome region}
\description{
This method draw specific genome region.
}
\examples{
genome.ver  <- 'mm10'
zoom.power  <- 1
quartz.bam.files   <- c(
  "~/Dropbox_RIKEN/Public_ACCCBiT/Data/Quartz-Seq/bam/Quartz_01.bam",
  "~/Dropbox_RIKEN/Public_ACCCBiT/Data/Quartz-Seq/bam/Quartz_02.bam"
)

gb.quartz <- orenogb2$new(
  genome.ver  = genome.ver,
  zoom.power  = zoom.power,
  bam.files   = quartz.bam.files
)

# Coordination
gb.quartz$chr      <- "chr17"
gb.quartz$start.bp <- 35492880
gb.quartz$end.bp   <- 35526079
gb.quartz$plotgb()
}
\keyword{plot}

