% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotgb}
\alias{plotgb}
\title{draw a genome region}
\description{
This method draw specific genome region.
}
\examples{
q01.bam <- system.file("extdata", "Quartz_01.Pou5f1.bam", package = "orenogb2")
q02.bam <- system.file("extdata", "Quartz_02.Pou5f1.bam", package = "orenogb2")
genome.ver <- 'mm10'
zoom.power <- 1
quartz.bam.files <- c(q01.bam, q02.bam)

gb.quartz <- orenogb2$new(
  genome.ver = genome.ver,
  zoom.power = zoom.power,
  bam.files  = quartz.bam.files
)

# Coordination
gb.quartz$chr      <- "chr17"
gb.quartz$start.bp <- 35492880
gb.quartz$end.bp   <- 35526079

png("Pou5f1.png", height = 600, width = 600)
gb.quartz$plotgb()
dev.off()

gb.quartz$zoom.power <- 1/200
png("Pou5f1.zoom.png", height = 600, width = 600)
gb.quartz$plotgb()
dev.off()
}
\keyword{plot}

